+++ abci/app/v1/common.go
@@ -39,6 +39,8 @@ var (
 	idpListKeyBytes                               = []byte("IdPList")
 	allNamespaceKeyBytes                          = []byte("AllNamespace")
 	servicePriceMinEffectiveDatetimeDelayKeyBytes = []byte("ServicePriceMinEffectiveDatetimeDelay")
+	allowedIALListKeyBytes                        = []byte("AllowedIALList")
+	allowedAALListKeyBytes                        = []byte("AllowedAALList")
 )
 
 const (